Update customized built-in elements is="" interaction
This follows the DOM changes in https://github.com/whatwg/dom/pull/566,
and is part of fixing https://github.com/whatwg/html/issues/3402.

The normative changes are to the serialization algorithm, which now
writes out the element's is value as its is="" attribute, if no actual
is="" attribute is present. The rest of the changes are to introductory
text about customized built-in elements.