Make document.open() not reset the encoding to UTF-8
This change was made based on a discussion in https://bugzilla.mozilla.org/show_bug.cgi?id=380383. However, as noted in that bug Firefox never shipped with this change and neither did Chrome or Safari.

(Note that the initial change was to reset to UTF-16 in 37c4d16c18fc33eeb40f820ca88209284761a25b. I could not find the rationale for changing that to UTF-8 in 011b5e8403a557071961eb3f54bcee8dfa01138c.)

Edge is the only browser that exhibits this behavior.

Tests: https://github.com/w3c/web-platform-tests/pull/10756.