Run importScripts() fetching steps in parallel
Fixes #3775.

As noted there in more detail, this doesn't change much of the
semantics. Execution order is still synchronous, and errors are still
thrown in the same deterministic fashion for both network failures and
script-running failures. Indeed, Gecko already implements these
semantics.

It's likely that at the time this specification was originally written,
whether these fetches were in parallel or sequential was not observable,
at least by client-side code. However, these days, with the web
performance APIs and with service workers, it's very observable, and we
should be sure that parallel fetching is allowed.

For reference, here is a 2015-08-27 snapshot of this method, before any
of the script-loading refactorings and changes due to module scripts:

https://html.spec.whatwg.org/commit-snapshots/c9e804f04d03a0658bfa689cb0f368a4d2e37936/#dom-workerglobalscope-importscripts

You can see that it was sequential even then.