Consistently use no quotes in spec citations
This change normalizes the markup for spec citations to consistently
omit quotation marks.

There are 500+ instances of spec citations in the source, more than 90%
which of which use markup in the form `<ref spec=CSSOMVIEW>`. But there
are about 48 instances that were added to the spec instead in the form
`<ref spec="CSP">`, with quotation marks around this spec abbreviation.

This change drops the quotation marks from those 48 instances.