Return empty on "do not run" for running a script
Previously, as of ad3dda51552aeeeea60540c7f1b513fe45b486a5, the
algorithm returned NormalCompletion(undefined) when the script was not
run.

But the service worker script will need to distinguish between when the
script was run vs. not, for deciding whether it can proceed with
installation. The counterpart service worker spec change is:
https://github.com/w3c/ServiceWorker/pull/1346.