Clarify and tweak <embed> processing model
This rewrites the <embed> processing model in a more modern style, along
the way clarifying a few things and making some minor normative changes.

* Fixes #3876 by colocating the text about notification of there being
  no plugin into the main algorithm, where it is more easily noticed.
* Changes to not fire load events when no plugin is loaded.
* Changes the SVG browsing context to get unloaded when a non-SVG is
  being displayed.

With regard to these normative changes, browsers behave inconsistently.
Indeed, they behave inconsistently in general with regard to lots of
features of <embed> and <object>; see
https://github.com/whatwg/html/labels/topic%3A%20embed%20and%20object.
We leave a more comprehensive overhaul, including with a testing story,
for later work. This revision just provides a basis for that future
work.