Add requestAnimationFrame() in workers
Fixes #3587. Based on the proposal at
https://github.com/junov/OffscreenCanvasAnimation/blob/2e0546417d4f45d194270a67a1cdf303f2e0ef88/OffscreenCanvasAnimation.md.

This introduces a new AnimationFrameProvider mixin interface to contain
requestAnimationFrame() and cancelAnimationFrame(). It's included in
Window and DedicatedWorkerGlobalScope. In addition to generalizing them
to work on both objects, the two methods were rewritten editorially to
be based on Infra primitives.

This change also introduces the concept of updating the rendering of
worker event loops.