Integrate with requestIdleCallback
This adds a hook to the event loop's processing model to define when the start
of an idle period should begin. This deals with the issue in
w3c/requestidlecallback#70 and calls directly into the "start an idle period"
algorithm, instead of that algorithm having to having to spin the event loop.

This change depends on the PR in w3c/requestidlecallback#75.