Do not run FinishDynamicImport steps in parallel
These steps do a variety of things that only make sense back on the main thread.
The intention here was that fetching not block the main thread. However, that is
already the case even without going in parallel, given the "async algorithm"
framework of the script-fetching algorithms. (That framework is admittedly a bit
unclear; see https://github.com/whatwg/infra/issues/181.)