[WIP] JSON module support
This patch provides JSON modules as a single default export, with
parse errors checked before instantiating the module graph.

Note, editorially, it's unclear whether JSON modules should be
considered a type of "module script", with a settings object, fetch
options, base URL, etc or not. This patch considers them "module
scripts", but leaves those record fields unset (as they are unused).

This patch is based on
https://github.com/tc39/proposal-javascript-standard-library/pull/44
which hasn't landed yet, so the references are a bit awkward, and
this patch should not land until that one does.

Closes #4315