Editorial: move navigation document initialization
Previously, it was indented into the middle of "process a navigate
response", which made that algorithm really hard to read. There was also
a standalone "set the Document's address" algorithm which was never
called, but said it must be performed at the appropriate time.

This moves the algorithm out to be a sibling of "process a navigate
response", with appropriate inputs, and includes the step to set the URL
directly in the algorithm.

Closes #3327.