rendering: Update the conditions where an image creates a replaced box.
This updates the spec for two changes I've done / doing to Firefox, which also
match WebKit and Blink:

 * https://bugzilla.mozilla.org/show_bug.cgi?id=1196668
 * https://bugzilla.mozilla.org/show_bug.cgi?id=1472637

In particular, the two cases that need to work are:

 * `<img>` (without `alt` text) creates a replaced box, regardless of specified
   dimensions or what not.

 * `<img src="foo" alt="bar">` creates a replaced box while the image is
   loading, rather than an inline.

I added tests for the first in https://phabricator.services.mozilla.com/D11194,
since it matches all implementations.

I don't know how to add a test for the second that isn't racy.