Remove compound microtasks
Per https://lists.w3.org/Archives/Public/www-archive/2016Oct/0001.html the rationale for having this concept is the ability for some API to spin the event loop. This used to be the case with showModalDialog(), but that API got removed.

Implementations also do not have this concept and treat this as a regular microtoask that just happens to do quite a few things (going through mutation observer callbacks and dispatching slotchange events).