Defer sandbox and feature policy initialization.
The initialization steps for a new browsing context attempted to
initialize the new document's sandbox and feature policy before the
browsing context was made the nested browsing context of its container.
This fixes that by extracting a set of steps into a new algorithm which
can be run after the browsing context tree is correct.

Fixes: #4444