Editorial: uniformize issue references
This makes all issue references use links of the form "issue #NNN", or
"org/repo issue #NNN". This was the prevailing convention, but there
were a few exceptions.

Behind the scenes, this also removes the one usage of the "big-issue"
CSS class, replacing it with "XXX" (for which it was a stylistic
synonym, until
https://github.com/whatwg/whatwg.org/commit/36b8b4571e480ea54938052f7335fc0b9f3f0cee
removed the "big-issue" styles). It also changes the source from using
class="&#x0058;&#x0058;&#x0058;" to using class="XXX"; the former was
a workaround for an overzealous lint script, which was fixed in
https://github.com/whatwg/html-build/commit/3c51d80616fcd3d5a125a23bb63b5051399fe35d.

Finally, this removes a big commented-out chunk of canvas spec text for
fillVerticalText() and strokeVerticalText() methods, which were never
implemented.