Custom elements: add element internals
This provides the ElementInternals interface, which can be obtained for
custom elements via the element.attachInternals() method. For now
ElementInternals is empty, but it will gain members in #4383.

This also adds the ability for custom elements to set the
disabledFeatures static property, to disable element internals and
shadow DOM. Some DOM-side infrastructure work there is necessary in
https://github.com/whatwg/dom/pull/760.

Tests:
- https://github.com/web-platform-tests/wpt/pull/15123
- https://github.com/web-platform-tests/wpt/pull/15516
- https://github.com/web-platform-tests/wpt/pull/16853

Fixes https://github.com/w3c/webcomponents/issues/758.