Restore override URL concept for javascript: URLs
This concept was deleted in 6440ccae73 along with "overridden overload",
but it was discovered in #3989 that correct handling for javascript:
URLs still depend on this concept. The spec around this area has changed
quite a bit since then due to intermediate cleanups such as e13762fb62,
so this is not a simple revert.

Also, fix ungrammatical "append URL URL list" that was accidentally
introduced in e13762fb62.

Addresses some concern raised in #3989.