Editorial: refactor module script fetching algorithms
This creates separate top-level algorithms for each entry point into
module script graph fetching in the spec:

* Fetch an external module script graph
* Fetch an import() module script graph
* Fetch a modulepreload module script graph
* Fetch an inline module script graph
* Fetch a module worker script graph

Previously, for several of these entry points, the parts of the spec
that implemented them would directly use the lower-level algorithms
(such as "fetch a single module script" or "internal module script graph
fetching procedure"). Now, those lower-level algorithms are entirely
encapsulated within the "Fetching scripts" section.

This also introduces a diagram showing the relationship of the
algorithms to each other, and rearranges some of the lower-level
algorithms for an easier reading flow.