Rendering: further details/summary cleanup
As is evident from https://bugs.chromium.org/p/chromium/issues/detail?id=973074 the current description still had too much from the old description and led to some inconsistencies in implementations. This makes it clear the slots have no special styling (by default) and it's the first summary element that is special, not the first slot.