Fix document URL setting for javascript: URLs
A series of changes starting with 6440ccae73 which removed the override
URL concept as well as intermediate cleanups such as e13762fb62 eroded
the correctness of javascript: document URLs. Now it seems to be the
case that fixing this typo restores the correct handling for such URLs.

Addresses some concerns raised in #3989.