Gate `javascript:` navigation on sandboxing flags.
As `javascript:` URLs execute in the target browsing context, it's
somewhat surprising that they can be executed from a sandboxed document
if that document is allowed to create unsandboxed popups.

This patch gates `javascript:` URL execution on the source document's
sandboxing flags, which is more in line with developer expectations.