Revamp the user activation model
This replaces the previous concept of "triggered by user activation"
with a more nuanced model, introducing new concepts and a more detailed
processing model. This aligns with Chromium's "user activation v2" work,
which gained consensus during the discussion at TPAC [1]. (The previous
specification did not align with any implementation.)

This serves mostly as a foundational improvement. Interop work remains
to be done on individual APIs and how they use user activation, as well
as the exact set of events which count as user activation. Those issues
are tracked in [2] and elsewhere.

Fixes #1903. Fixes #3859.

[1]: https://docs.google.com/document/d/1gHxQMdXGX4UjjoPXi0c1vhwYregQzWNEHqgbydopCoo/edit#heading=h.9hw95kikso76
[2]: https://github.com/whatwg/html/labels/topic%3A%20user%20activation