Meta: keep the fragment when clicking the Multipage or Dev links
For people who follow a link to single page (with a fragment) but
prefer to load the multipage version, using the link at the top
would remove the fragment. This keeps the fragment when clicking
the link, and `link-fixup.js` will do the necessary redirect when
the multipage page is loaded.