Clarify that <input> where step does not apply has no allowed value step
If a content attribute does not apply to an <input> element, then the
user agent must "ignore the attribute" – essentially treating it as if
the attribute does not exist. Currently, if the "step" attribute does
not apply, then the definition breaks down, as there is no default step
defined for the element.

Fix this by explicitly defining that no allowed value step exists for
non-steppable <input> elements.