Fix confusable same site examples
The results shown in the example table for "same site" and "schemelessly
same site" algorithms are based on assuming that "suffix.example" is a
public suffix. Since it actually isn't, this causes confusion.

This commit replaces "suffix.example" with "wildlife.museum" which is a
valid public suffix from the Public Suffix List
(https://publicsuffix.org). It also explicitly states all of the public
suffixes involved in the examples.

Closes #5316.