Editorial: clean up ambiguous usages of string's "length" and "offset"
This change is a follow-up to a9ef15d in aligning the HTML spec with the
string changes in Infra (see whatwg/infra#291). In particular, it
changes existing usages of "length" which didn't link to either of the
Infra definitions, and it updates usages of offsets to have them refer
to the string's code unit representation.

As part of updating the usage of offsets in form control algorithms,
this change also removes text which is no longer normative or turns it
into a note.