Make document's character encoding reflect byte-order marks
This change fixes a bug where document's character encoding was set to
the return value of the encoding sniffing algorithm rather than to the
actual encoding used, which differed when the stream started with a
byte-order mark. This change incorporates BOM sniffing into the encoding
sniffing algorithm, ensuring both encodings are identical.

Closes #1077.