More explicitly associate agents and event loops
This adds a specific definition for "an agent's event loop", as opposed
to "event loop" generally, and uses it where appropriate. In a few
cases, this involves moving from "object's event loop" to "object's
relevant agent's event loop".

Note that for worker agents, event loop creation and association with
the agent is explicit. For window agents there is no corresponding text.
This can be settled as part of #5411.