Change "scripting is disabled" to be per-global
Previously it was per-browsing context. This caused a number of
problems when used from workers; it went through the "responsible
document" indirection which did not exist for service workers and was
racy in other cases.

In theory this might be cleaner if it were per agent, i.e. if agents
maintained internal consistency about disabling or enabling scripting.
In practice user agents seem to have a single user-agent-wide switch,
and at least some of them wire that to a per-global boolean, so we just
use per-global as it fits most cleanly into existing architecture and
matches at least some implementation internals.

Closes #1580. Helps with #5422.