Change definition of string equality to use 'is'/'identical to' from Infra
Fix #5067

Adds the (new) definition of string equality using terms 'is' or 'identical to' from Infra, replacing the term 'case-sensitive'. The phrase "case sensitive" is retained in the text where it provides clarity (particularly in sections with names like "case-sensitivity of XXX").

I may need to go back and define an inverse term: there are a number of places where "case-sensitive" was used with "unordered set of unique space-separated tokens". A number of these remain in the commit.