Redo top-level origin and add top-level creation URL
An environment's top-level origin is null during the initial top-level navigation (before the response arrives) and otherwise represents the origin of the top-level document. It is currently implementation-defined for non-dedicated workers, but hopefully that can be sorted soon.

An environment's top-level creation URL is the URL of the top-level document. It is null for workers as they do not need the concept.

Needed for https://github.com/whatwg/fetch/pull/943 and #5558.

Co-authored-by: Anne van Kesteren <annevk@annevk.nl>