Tweak "is unloading" in navigation
This replaces vague text about "if the prompt to unload algorithm is
being run" and the like for "unload" with an explicit check of the
"unload counter". The "unload counter" is a rename of the
"ignore-opens-during-unload counter", which was already used for similar
explicit checks in document.open().

This results in the same check being done everywhere, which appears to
match at least some implementations better. Previously, for example,
"traverse the history by a delta" did not check for beforeunload, only
unload.