Set dfn-type on definitions
Based on a specialized definitions parser for the HTML spec built to improve on the existing scraping done in shepherd.

This particular patch doesn't cover all the identified dfn-type that could be set, only the ones that were easy to treat (i.e. where the data-x and the dfn occur on the same line).

The setting of dfn-type and dfn-for is based on the WebIDL defined in the spec