Define navigation parameter: "replacement enabled"
Many invocations of the "navigate" algorithm (along with the algorithm
itself) refer to the "replacement enabled" flag, but this flag is not
listed among the possible input to the algorithm.

Extend the input to the algorithm with the flag, and (due to the
prevalence of this usage) make this the canonical definition of the
term.