Specify sequence of navigation failure checks
The clause "if any of the following are true" does not indicate the
intended sequence by which the subsequent conditions must be evaluated.
Conforming implementations are therefore free to perform the checks in
any order. However, the sequence is observable by web content because
both the Content-Security Policy (CSP) check and the Cross-Origin
Embedder Policy (COEP) check may trigger reporting events.

The current implementation status is as follows (neither Safari nor
Firefox implement COEP reporting at this time):

- Chrome 86:   CSP, then COEP, then X-Frame-Options
- Firefox 80:  CSP, then X-Frame-Options
- Safari 13.1: CSP, then X-Frame-Options

Standardize the sequence according to the state of the implementations.