Remove quotes UA rendering section.
The CSS spec[1] has been updated to use 'auto' as the initial value for
quotes, as resolved in [2]. The 'auto' behavior is what the HTML spec
has been trying to express. Instead, let the HTML elements have the
initial value for the quotes property and let CSS specify what 'auto'
means.

[1] https://drafts.csswg.org/css-content/#quotes-property
[2] https://github.com/w3c/csswg-drafts/issues/4074