Allow autocomplete="username" with <input type=email>
This change updates the normative document-conformance requirements for
the "autocomplete" attribute to allow documents to have an "input"
element such as the following:

<input name="u" type="email" autocomplete="username">

That is, this change allows an "input" element whose type is "email" to
have an "autocomplete" attribute whose value is "username".

Otherwise, without this change, the spec does not allow the "username"
value for the "autocomplete" when the input type is "email".

This change also takes what had been called the "E-mail" control group,
and renames it to the "Username" control group.

Fixes #5873. See also https://github.com/validator/validator/issues/983.