Make beforeunload not affect 'salvageable' & firing unload event optional
Currently firing beforeunload on a document will set its 'salvageable'
bit to false, which is not true in practice in some user agents.

Also, firing the 'unload' event should be optional if the user agent
decides that it will keep the document alive in a session history entry,
to be used later.