Meta: set dfn types; switch to (Bikeshed) short syntax
This change sets dfn type attributes on a large number of dfn and other
elements throughout the spec — generated using a specialized definitions
parser for the HTML spec built to improve on the existing scraping done
in Shepherd.

This particular patch doesn’t set all the potential dfn types that could
be set; instead it only sets those that were easiest to determine (those
where the data-x attribute and dfn start tag occur on the same line).

The setting of dfn type attributes and “for” attributes is based on the
corresponding WebIDL defined in the spec.

Note that this change also switches the spec to using the Bikeshed short
syntax for the attribute names; specifically:

* rather than using data-dfn-type=element, etc., attribute names, it
  uses attributes literally named "element", etc.

* rather than the attribute name data-dfn-for it uses an attribute
  literally named "for"

* replaces all data-export="" and data-noexport="" attributes with
  value-less attributes literally named "export" and "noexport"

Additionally, the change drops "export" from any dfn that has a dfn type
attribute, and switches data-dfn-type="dfn" disambiguators to "export"
instead. It also cleans up some of the markup in the dependencies
section, which shouldn't be parsed anyway.