Meta: properly export global attributes and their values
Closes #5417. Previously many of these were marked up as element-attr,
but without a for="", which is invalid according to the Bikeshed
linking contract. dir="" was especially messed up, with its *values*
marked up as element-attr, but the attribute itself not marked up at
all.

This uses for="global" for any-namespace elements and for="html-global"
for HTML elements.

This includes some other related fixes such as:

* Unifying the inconsistent definition location for these attributes.
  Sometimes they would be defined in their section headers; sometimes
  separately. Now they are all defined separately.

* Marking up some of the related IDL attributes.

* Delegating definitions of xml:space, xml:lang, and the style IDL
  attribute to other specifications, instead of re-defining them in
  HTML.