Restore old anchor for input Email state
This change adds an additional anchor to the dfn/heading for the
Email state of the input element.

Otherwise, without this change, we have broken links at any sites which
are were linking to the old "e-mail-state-(type=email)" anchor — that
is, with "e-mail" rather than "email". (That changed when we globally
replaced "e-mail" with "email" throughout the source.)