Unwrap the heading for the meta element section
This change unwraps the heading for the meta element section from <100
characters back to >100 characters.

Otherwise, without this change, the .pre-process-tag-omission.pl script we
use for generating the **Tag omission in text/html** subsections for every
element section doesn’t work as expected for the meta element section.

The reason is, that script is written to read in the source line-by-line
and so doesn’t expect any headings to break across multiple lines.