Fixup parenthesis placement in previous commit
f76d3c5cc1287c372670e4267c439da340e17e24 placed the parenthesis
incorrectly.

Also note that its commit message was incorrect in an important aspect:
public-suffix roots are *not* considered same-site to their subdomains.
Instead the relevant change was that public-suffix roots need to be
compute to the same site as versions of themselves with different ports.