Change element.click() to fire a PointerEvent
In the effort to change click event from a MouseEvent to a PointerEvent,
we need to change HTMLElement.click() to fire a "click" PointerEvent at
the target instead of a "click" MouseEvent.

Note: Given that PointerEvent inherits from MouseEvent, going forward
HTMLElement.click() will keep firing a MouseEvent at the target.