Fix regression in focus chain computation
This regressed in 90a60b2a0dc740b8b0093b07ca0a41e70ba8d83a, which
accidentally removed the step which lets the recursion continue by
setting the current object to the appropriate "owner" document. Even
though control group owner objects are removed, we still need to
traverse up to the document in focus chain computation.

While here, this rewrites the algorithm to use a loop instead of a goto
for clarity.

Closes #6161.