Fix update a style block to match reality and CSSWG resolution.
No browser passed the test-case that was linked from the comment.

The algorithm was also inconsistent, as it looked at whether
the element was in a shadow tree or in the document tree, but it was
only specified to be re-run if the element becomes connected or
disconnected.

The CSSWG discussed this in
https://github.com/w3c/csswg-drafts/issues/3096#issuecomment-420719482
and there are WPTs for this.

This also matches closer the definition of <link rel="stylesheet"> which
does use connectedness (though it uses "browsing-context connected",
which is a bit different):
https://html.spec.whatwg.org/#link-type-stylesheet