Suggest using :focus-visible instead of :focus to override the focus ring
Currently the spec suggests authors to use :focus to hide an outline when an element is focused,
but actually :focus-visible should be the pseudo-class used for that.

This patch also modifies the example to use a textarea instead of a link,
as links not always show a focus ring when focused, while textarea elements always do it.