Move steps into "URL and history update steps"
These three steps were previously only performed for history.pushState()/history.replaceState(), but they need to be performed for all callers of the URL and history update steps, the other one being document.open().

For the serializedData steps, this is editorial, as document.open() does not pass serializedData. But for updating the latest entry, it's a bug fix, as otherwise things get strangely out of sync.