Add hooks for other specifications to observe navigation
This change is motivated by specifications, notably WebDriver-BiDi,
that want to observe the progress of navigations. Currently
specifications that want to do this effectively monkeypatch HTML by
writing "when <some stageis reached>, run the following:". The aim
here is to provide more definite integration points.

Implementation wise, this adds a single new concept; a navigation id
which is a unique id generated for each navigation. This enables
consumers to tell when events are from the same navigation.

On top of that, a navigation status struct is defined, with an id, a
URL and a status. This is used as a uniform interface for
communicating the navigation progress through all the integration
points, even though not all fields are useful in all cases.

The actual integration points added are:

* Navigation started
* Navigation failed
* Download started
* DOM load complete
* Load complete

In addition the navigate algorithm itself returns a navigation status,
so that external specifications which initiate a navigation can get
the associated navigation id (or find out that the navigation
completed in the case of a fragment navigation).