Add WebCodecs VideoFrame as CanvasImageSource.
Adding VideoFrame as a CanvasImageSource allows drawImage(),
createImageBitmap(), and texImage() to interoperate with VideoFrames.
VideoFrames are effectively the same as the existing point-in-time
capture done for HTMLVideoElement on each of these interfaces.

This adds a non-normative reference on the WebCodecs spec for the
VideoFrame interface and associated properties to accomplish this.

Bug: https://github.com/w3c/webcodecs/issues/158