Improve normalization of the resulting URL record of protocol handlers
Currently, it requires the URL's scheme to be `https` which is a bit too
strict compared to being "same origin as the secure context where the
normalization happens". For example, it does not allow developers to
quickly test things using a local http server. It also does not correspond
to what is implemented in browsers. The condition is relaxed to being an
HTTP(S) scheme which is good enough given the mentioned requirement of
being "same origin as the secure context". An assertion is added to make
clear this requirement implies that the url is potentially trustworthy.
The assertion is handy to see if the normalization algorithm is reused by
implementers in a different context.

Closes #5787