Always replace the initial about:blank Document
Previously, the caller of the navigate algorithm would (usually) check for the browsing context still being on the initial about:blank Document, and if so, switch its history handling to "replace". However, a few call sites missed this: e.g., following hyperlinks or window.open() when they were targeted at an existing browsing context.

This change instead centralizes the the conversion of "default" navigations into "replace" navigations for the initial about:blank Document.