Fixes and cleanups to "current document readiness"
* Explicitly call out to "update the current document readiness" when we want the side effects to happen.

* Do not fire readystatechange if the value does not actually change.

* Update the insertion point before firing readystatechange.

* Make parser-less documents not be a special case, because most of them end up calling "stop parsing" before any script runs anyway.

* Move everything related to current document readiness into the "Reporting document loading status" section.

Most of these are not observable, because script cannot run during the times when these values are changing. The exception is that we now update the insertion point before firing readystatechange, whereas the previous not-very-precise phrasing seemed to imply that you'd update the insertion point afterward.

The new order matches implementations, as seen from the test at https://github.com/web-platform-tests/wpt/pull/28799.