Support automatic user prompt handing in WebDriver BiDi
This reads the return value from the WebDriver BiDi user prompt opened
algorithm and uses it to decide whether to show the prompt, or skip
showing it but act as if it had been accepted or canceled.

At present the prompt opened function can return three values: "none",
which causes the prompt to be displayed, "accept", which acts as if
the prompt was accepted, and "dismiss" which acts as if the prompt was
rejected. The WebDriver side is responsible for ensuring that only
valid responses are sent for each prompt type (e.g. for an alert there
isn't a "dismiss" option).