Made normative and non-normative sections consistent
Chrome, Safari and Firefox are now consistent in their handling of all-named elements, so removed the 'in some browsers' comments.

Added the missing elements from the normative list to the non-normative list, with the exception of <frame> and <frameset> as they aren't conforming.

The result from the https://software.hixie.ch/utilities/js/live-dom-viewer/?saved=2837 test (in a comment in 'source') is identical in all major browsers, and demonstrates they use the same set of all-named elements.